[WOOR-34] refactor: inviteCode 패키지 이동
