+++ src/main/java/com/musseukpeople/woorimap/couple/application/CodeGenerator.java
@@ -1,4 +1,4 @@
-package com.musseukpeople.woorimap.invitecode.application;
+package com.musseukpeople.woorimap.couple.application;
 
 public interface CodeGenerator {
