+++ src/main/java/com/musseukpeople/woorimap/couple/application/dto/response/InviteCodeResponse.java
@@ -1,4 +1,4 @@
-package com.musseukpeople.woorimap.invitecode.application.dto.response;
+package com.musseukpeople.woorimap.couple.application.dto.response;
 
 import lombok.AccessLevel;
 import lombok.Getter;
