+++ src/main/java/com/musseukpeople/woorimap/couple/domain/InviteCodeRepository.java
@@ -1,8 +1,8 @@
-package com.musseukpeople.woorimap.invitecode.domain;
+package com.musseukpeople.woorimap.couple.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.musseukpeople.woorimap.invitecode.infrastructure.InviteCodQueryRepository;
+import com.musseukpeople.woorimap.couple.infrastructure.InviteCodQueryRepository;
 
 public interface InviteCodeRepository extends JpaRepository<InviteCode, String>, InviteCodQueryRepository {
 }
