+++ src/main/java/com/musseukpeople/woorimap/couple/infrastructure/InviteCodQueryRepositoryImpl.java
@@ -1,11 +1,11 @@
-package com.musseukpeople.woorimap.invitecode.infrastructure;
+package com.musseukpeople.woorimap.couple.infrastructure;
 
 import static com.musseukpeople.woorimap.invitecode.domain.QInviteCode.*;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
 
-import com.musseukpeople.woorimap.invitecode.domain.InviteCode;
+import com.musseukpeople.woorimap.couple.domain.InviteCode;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
