+++ src/test/java/com/musseukpeople/woorimap/couple/infrastructure/InviteCodQueryRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.musseukpeople.woorimap.invitecode.infrastructure;
+package com.musseukpeople.woorimap.couple.infrastructure;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
@@ -11,8 +11,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import com.musseukpeople.woorimap.invitecode.domain.InviteCode;
-import com.musseukpeople.woorimap.invitecode.domain.InviteCodeRepository;
+import com.musseukpeople.woorimap.couple.domain.InviteCode;
+import com.musseukpeople.woorimap.couple.domain.InviteCodeRepository;
 import com.musseukpeople.woorimap.util.RepositoryTest;
 
 class InviteCodQueryRepositoryTest extends RepositoryTest {
